# my-music-go/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AddSongToPlaylistRequestDTO:
    properties:
      song_id:
        type: integer
    required:
    - song_id
    type: object
...
  models.CreatePlaylistRequestDTO:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  models.PaginatedResponseDTO:
    properties:
      list:
        description: 使用 interface{} 来适应任何类型的列表数据
      total:
        type: integer
    type: object
  models.PlaylistDetailDTO:
    properties:
      cover:
        type: string
      created_at:
        type: string
      creator_name:
        type: string
      description:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      name:
        type: string
      song_count:
        type: integer
      songs:
        allOf:
        - $ref: '#/definitions/models.PaginatedResponseDTO'
        description: 复用分页结构
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.PlaylistInfoDTO:
    properties:
      cover:
        type: string
      created_at:
        type: string
      creator_name:
        type: string
      description:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      name:
        type: string
      song_count:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.UpdatePlaylistRequestDTO:
    properties:
      cover:
        type: string
      description:
        type: string
      is_public:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: your.email@example.com
    name: Elaine
    url: https://github.com/Garhlz
  description: 这是一个使用 Go 语言和 Gin 框架开发的音乐播放器后端 API.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Garhlz
  title: Elaine 的音乐播放器 API
  version: "1.0"
paths:
...
  /me/playlists:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponseDTO'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/models.PlaylistInfoDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取我的歌单列表
      tags:
      - 歌单 (Playlist)
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建歌单请求
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CreatePlaylistRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: '{"id": 1, "message": "创建成功"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建我的新歌单
      tags:
      - 歌单 (Playlist)
  /me/playlists/{playlistId}:
    delete:
      parameters:
      - description: 歌单 ID
        in: path
        name: playlistId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "删除成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除我的歌单
      tags:
      - 歌单 (Playlist)
    put:
      consumes:
      - application/json
      parameters:
      - description: 歌单 ID
        in: path
        name: playlistId
        required: true
        type: integer
      - description: 更新歌单请求
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePlaylistRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "更新成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新我的歌单信息
      tags:
      - 歌单 (Playlist)
  /me/playlists/{playlistId}/songs:
    post:
      consumes:
      - application/json
      parameters:
      - description: 歌单 ID
        in: path
        name: playlistId
        required: true
        type: integer
      - description: 添加歌曲请求
        in: body
        name: addSongRequest
        required: true
        schema:
          $ref: '#/definitions/models.AddSongToPlaylistRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: '{"message": "添加成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 向我的歌单添加歌曲
      tags:
      - 歌单 (Playlist)
  /me/playlists/{playlistId}/songs/{songId}:
    delete:
      parameters:
      - description: 歌单 ID
        in: path
        name: playlistId
        required: true
        type: integer
      - description: 歌曲 ID
        in: path
        name: songId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "移除成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 从我的歌单移除歌曲
      tags:
      - 歌单 (Playlist)
...
  /playlists/{playlistId}:
    get:
      parameters:
      - description: 歌单 ID
        in: path
        name: playlistId
        required: true
        type: integer
      - description: 歌曲列表页码
        in: query
        name: page
        type: integer
      - description: 歌曲列表每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlaylistDetailDTO'
      security:
      - BearerAuth: []
      summary: 获取歌单详情
      tags:
      - 歌单 (Playlist)
...
  /users/{id}/playlists:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponseDTO'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/models.PlaylistInfoDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取指定用户的歌单列表
      tags:
      - 歌单 (Playlist)
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"